add prettier for solidity

+++ contracts/Auction.sol
@@ -7,13 +7,18 @@ pragma solidity ^0.6.2;
 contract Auction {
   address public factory;
   address public seller;
-  uint public sellerDeposit;
-  uint public tokenAmount;
+  uint256 public sellerDeposit;
+  uint256 public tokenAmount;
   address public tokenContractAddress;
-  uint public startDateTime;
-  uint public endDateTime;
+  uint256 public startDateTime;
+  uint256 public endDateTime;
 
-  constructor(uint _tokenAmount, address _tokenContractAddress, uint _startDateTime, uint _endDateTime) public {
+  constructor(
+    uint256 _tokenAmount,
+    address _tokenContractAddress,
+    uint256 _startDateTime,
+    uint256 _endDateTime
+  ) public {
     factory = msg.sender;
     tokenAmount = _tokenAmount;
     tokenContractAddress = _tokenContractAddress;
@@ -25,6 +30,7 @@ contract Auction {
     require(msg.sender == factory);
     seller = _seller;
   }
+
   function receiveSellerDeposit() external payable {
     require(msg.sender == seller);
     sellerDeposit = msg.value;
